# api-gateway/src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: Employee Management System API
  description: |
    Comprehensive Employee Management System API Gateway
    
    ## Overview
    This microservices-based Employee Management System provides complete functionality for managing employees, departments, and authentication through a unified API Gateway.
    
    ## Authentication
    - **JWT Bearer Token**: All protected endpoints require a valid JWT token
    - **Test Credentials**: Use POST /api/auth/login with:
      - Username: `admin`
      - Password: `admin123`
    
    ## Services
    - **Auth Service**: User authentication and authorization
    - **Employee Service**: Employee and department management
    - **API Gateway**: Unified entry point (this documentation)
    
    ## Service URLs
    - API Gateway: http://localhost:8080
    - Auth Service: http://localhost:8082
    - Employee Service: http://localhost:8083
    
    ## Quick Start
    1. Authenticate via `/api/auth/login`
    2. Copy the JWT token from response
    3. Click "Authorize" button and enter: `Bearer YOUR_TOKEN`
    4. Explore protected endpoints
    
  version: 1.0.0
  contact:
    name: Employee Management System
    email: support@employeemgt.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: API Gateway (Primary)
  - url: http://localhost:8082
    description: Auth Service (Direct)
  - url: http://localhost:8083
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from /api/auth/login endpoint.
        Format: Bearer <token>
        
        **Quick Test**: Login with admin/admin123 to get a token
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Operation success status
          example: true
        message:
          type: string
          description: Human-readable response message
          example: "Operation completed successfully"
        data:
          type: object
          description: Response payload data
          nullable: true
      required:
        - success
        - message

    # Authentication Schemas
...
    EmployeeRequest:
      type: object
      properties:
        employeeNumber:
          type: string
          description: Unique employee identifier (auto-generated if not provided)
          example: "EMP202410001"
          minLength: 3
          maxLength: 20
        firstName:
          type: string
          description: Employee first name
          example: "John"
          minLength: 2
          maxLength: 50
        lastName:
          type: string
          description: Employee last name
          example: "Doe"
          minLength: 2
          maxLength: 50
        email:
          type: string
          format: email
          description: Employee email address
          example: "john.doe@company.com"
        phoneNumber:
          type: string
          description: Employee phone number
          example: "+1234567890"
          minLength: 10
          maxLength: 15
        hireDate:
          type: string
          format: date
          description: Employee hire date
          example: "2024-01-15"
        jobTitle:
          type: string
          description: Employee job title
          example: "Software Engineer"
          minLength: 2
          maxLength: 100
        salary:
          type: number
          format: decimal
          description: Employee salary
          example: 75000.00
          minimum: 0
        status:
          type: string
          enum: [ACTIVE, INACTIVE, TERMINATED]
          description: Employee status
          example: "ACTIVE"
          default: "ACTIVE"
        departmentId:
          type: integer
          format: int64
          description: Department ID where employee belongs
          example: 1
        managerId:
          type: integer
          format: int64
          description: Manager ID (optional)
          example: 5
          nullable: true
        notes:
          type: string
          description: Additional notes about the employee
          example: "Experienced developer with expertise in Java and Spring"
          maxLength: 1000
      required:
        - firstName
        - lastName
        - email
        - hireDate
        - jobTitle
        - salary
        - departmentId
    EmployeeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Employee unique identifier
          example: 1
        employeeNumber:
          type: string
          description: Employee number
          example: "EMP202410001"
        firstName:
          type: string
          description: Employee first name
          example: "John"
        lastName:
          type: string
          description: Employee last name
          example: "Doe"
        fullName:
          type: string
          description: Employee full name
          example: "John Doe"
        email:
          type: string
          description: Employee email
          example: "john.doe@company.com"
        phoneNumber:
          type: string
          description: Employee phone number
          example: "+1234567890"
        hireDate:
          type: string
          format: date
          description: Employee hire date
          example: "2024-01-15"
        jobTitle:
          type: string
          description: Employee job title
          example: "Software Engineer"
        salary:
          type: number
          format: decimal
          description: Employee salary
          example: 75000.00
        status:
          type: string
          enum: [ACTIVE, INACTIVE, TERMINATED]
          description: Employee status
          example: "ACTIVE"
        department:
          $ref: '#/components/schemas/DepartmentSummary'
        managerId:
          type: integer
          format: int64
          description: Manager ID
          example: 5
          nullable: true
        notes:
          type: string
          description: Additional notes
          example: "Experienced developer"
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2024-01-15T10:30:00"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-01-15T10:30:00"

    # Department Schemas
...
    DepartmentSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Department ID
          example: 1
        name:
          type: string
          description: Department name
          example: "Engineering"
        code:
          type: string
          description: Department code
          example: "ENG"

    # Filter Schemas  
...
    PaginatedApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                content:
                  type: array
                  items:
                    type: object
                  description: Page content items
                totalElements:
                  type: integer
                  format: int64
                  description: Total number of elements
                  example: 150
                totalPages:
                  type: integer
                  description: Total number of pages
                  example: 15
                currentPage:
                  type: integer
                  description: Current page number (0-based)
                  example: 0
                pageSize:
                  type: integer
                  description: Number of items per page
                  example: 10
                first:
                  type: boolean
                  description: Is this the first page
                  example: true
                last:
                  type: boolean
                  description: Is this the last page
                  example: false
paths:
...
  /api/employees:
    post:
      tags:
        - Employee Management
      summary: Create New Employee
      description: |
        Create a new employee record. Only ADMIN users can create employees.
        Employee number is auto-generated if not provided.
      operationId: createEmployee
      security:
        - bearerAuth: []
      requestBody:
        description: Employee information for creation
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeRequest'
            examples:
              new_employee:
                summary: Complete Employee Creation
                value:
                  employeeNumber: "EMP001"
                  firstName: "John"
                  lastName: "Doe"
                  email: "john.doe@company.com"
                  phoneNumber: "+1234567890"
                  hireDate: "2024-01-15"
                  jobTitle: "Software Engineer"
                  salary: 75000.00
                  status: "ACTIVE"
                  departmentId: 1
                  managerId: 5
                  notes: "Experienced developer with expertise in Java and Spring"
      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EmployeeResponse'
        '400':
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden - ADMIN role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/employees/{id}:
    get:
      tags:
        - Employee Management
      summary: Get Employee by ID (Admin Only)
      description: Retrieve a specific employee by ID. Only accessible by ADMIN users.
      operationId: getEmployeeById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Employee ID
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Employee retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EmployeeResponse'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden - ADMIN role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    put:
      tags:
        - Employee Management
      summary: Update Employee
      description: Update an existing employee record. Only ADMIN users can update employees.
      operationId: updateEmployee
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Employee ID
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        description: Updated employee information
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeRequest'
      responses:
        '200':
          description: Employee updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EmployeeResponse'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden - ADMIN role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags:
        - Employee Management
      summary: Delete Employee
      description: Delete an employee record. Only ADMIN users can delete employees.
      operationId: deleteEmployee
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Employee ID
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '204':
          description: Employee deleted successfully
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden - ADMIN role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/employees/all:
    get:
      tags:
        - Employee Management
      summary: Get All Employees with Filtering
      description: |
        Retrieve paginated list of employees with advanced filtering capabilities.
        Only accessible by ADMIN users.
        
        **Filter Options:**
        - Name, email, department, manager
        - Hire date range, salary range
        - Employee status
        - Pagination and sorting
      operationId: getAllEmployeesForAdmin
      security:
        - bearerAuth: []
      parameters:
        - name: firstName
          in: query
          description: Filter by first name
          schema:
            type: string
            example: "John"
        - name: lastName
          in: query
          description: Filter by last name
          schema:
            type: string
            example: "Doe"
        - name: email
          in: query
          description: Filter by email
          schema:
            type: string
            example: "john@company.com"
        - name: departmentId
          in: query
          description: Filter by department ID
          schema:
            type: integer
            format: int64
            example: 1
        - name: managerId
          in: query
          description: Filter by manager ID
          schema:
            type: integer
            format: int64
            example: 5
        - name: status
          in: query
          description: Filter by employee status
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, TERMINATED]
            example: "ACTIVE"
        - name: hireDateFrom
          in: query
          description: Filter by hire date from
          schema:
            type: string
            format: date
            example: "2024-01-01"
        - name: hireDateTo
          in: query
          description: Filter by hire date to
          schema:
            type: string
            format: date
            example: "2024-12-31"
        - name: salaryMin
          in: query
          description: Minimum salary filter
          schema:
            type: number
            format: decimal
            example: 50000.00
        - name: salaryMax
          in: query
          description: Maximum salary filter
          schema:
            type: number
            format: decimal
            example: 100000.00
        - name: page
          in: query
          description: Page number (0-based)
          schema:
            type: integer
            default: 0
            example: 0
        - name: size
          in: query
          description: Page size
          schema:
            type: integer
            default: 10
            example: 10
        - name: sort
          in: query
          description: Sort criteria (field,direction)
          schema:
            type: string
            default: "id,asc"
            example: "firstName,asc"
      responses:
        '200':
          description: Employees retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          content:
                            type: array
                            items:
                              $ref: '#/components/schemas/EmployeeResponse'
        '403':
          description: Forbidden - ADMIN role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/employees/department:
    get:
      tags:
        - Employee Management
      summary: Get Department Employees (Manager Only)
      description: |
        Retrieve employees belonging to the manager's department.
        Only accessible by MANAGER role users.
      operationId: getDepartmentEmployees
      security:
        - bearerAuth: []
      parameters:
        - name: firstName
          in: query
          description: Filter by first name
          schema:
            type: string
            example: "John"
        - name: lastName
          in: query
          description: Filter by last name
          schema:
            type: string
            example: "Doe"
        - name: page
          in: query
          description: Page number (0-based)
          schema:
            type: integer
            default: 0
            example: 0
        - name: size
          in: query
          description: Page size
          schema:
            type: integer
            default: 10
            example: 10
      responses:
        '200':
          description: Department employees retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden - MANAGER role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/employees/view:
    get:
      tags:
        - Employee Management
      summary: Get My Employee Profile (User Only)
      description: |
        Retrieve authenticated user's own employee profile information.
        Only accessible by USER role.
      operationId: getMyEmployeeProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Employee profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EmployeeResponse'
        '403':
          description: Forbidden - USER role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/employees/admin/test:
    get:
      tags:
        - Employee Management
      summary: Admin Test Endpoint
      description: Test endpoint for ADMIN users only
      operationId: employeeAdminTest
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Admin endpoint test successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                admin_test:
                  summary: Admin Test Response
                  value:
                    success: true
                    message: "Admin endpoint test successful"
                    data:
                      message: "Admin test endpoint working"
                      timestamp: "2025-10-29T12:00:00"
        '403':
          description: Forbidden - ADMIN role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # Department Management Endpoints
...